refactor(settings): extract connections pane into eh-settings-connections.js
Move the Health Auto Export panel (endpoint, token, last-push
diagnostics) and the hidden Apple Health metrics list out into their
own component. The settings shell is now ~25 lines that delegates
to <eh-settings-connections> + <eh-settings-cards>.

Sets up the tabbed shell that ships next: each pane drops into a tab
slot without further refactoring.

Refs #383.
